requirements.txt
pynput
PyAutoGui
pyperclip
python-gnupg